Added different environments
Added different environments - normal, harsh, desert, and rainforest. Each environment has a percentage modifier to use with the maximum energy an organism can have (for the 3 base types). Normal represents the numbers we started with passive - 20, herbivore - 50, and carnivore - 100. I figured out what percentages those were based on grid size and that became the starting values. Harsh is overall harsh for all. Desert is worse for passive plants and rainforest is worse for carnivores and better for plants. I figured we could add more environments in the future.