Modified environments, added reproduction rate mod, modified passive organism
I adjusted the values of the modifiers within the different environments. I also added reproduction rate mods for each of the three organisms which is now present in the genome and altered based on the environment. I adjusted the passive organisms to utilize this reproduction rate modifier for reproduction chance, and I also modified the passive organisms to have intermittent energy decay so they will not always persist in harsher environments.