Further cleaned up world started making environments more dynamic
Cleaned up world so that it only selects the type of environment and the rest of the methods should be in environment instead of world. Started making the environments more dynamic based on weather events. Initially started with rainforest by adding normal, rainy, and stormy weather which are often found in rainforests. So far a weather is chosen for a random amount of days and then a new weather event is then chosen and only passive organism max energy is effected so far.